backend/requirements.txt
fastapi[standard]==0.117.1
uvicorn==0.37.0
psycopg2-binary==2.9.11
sqlmodel==0.0.27
sqlalchemy==2.0.43
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1
qrcode[pil]==8.2
PyJWT[crypto]==2.10.1
argon2-cffi==25.1.0
email-validator==2.3.0
proxmoxer==2.2.0
requests==2.32.5
